refactor: corrige importaçoes e escalas de medidas
