+++ src/utils/index.ts
@@ -1,3 +1,8 @@
 export * from './device'
 export * from './breakpoints'
 export * from './buildDate'
+export * from './buildDate'
+export * from './buildMetrics'
+export * from './estatistica'
+export * from './extractDataToStatistic'
+export * from './buildSVG'
